+++ docs/fr_FR/changelog.md
@@ -18,6 +18,7 @@
 - Si le démon Local est en status NOK alors les présences locales sont mises a 0
 - Si une antenne n'a pas communiqué depuis plus d'une minute alors les présences de cette antenne sont mises à 0
 - Rajout de la possibilité de récupérer les nouvelles configurations sans mettre à jour le plugin
+- Rajout d'un mode passif ou actif pour le scan
 
 # 26/06/2019
 - Rajout du xiaomi cleargrass